Code improvment and reorganizing
modified the base KNOB V2 code to make it easier to modify what each key and dial does.

added code for a dial that both scrolls pages and uses the keys to switch and delete tabs.

finally renamed one of the files to correct the error when opening in the arduino IDE